Contributed by mkarasek
roll.count now starts at 0 instead of 1.

New roll option: u(nformatted). Same output as e, but without the formatting (i.e. not wrapped in a span tag) so it can be embedded in a tooltip.

p(lain) roll option renamed to r(esult).

Parameter for t option is now evaluated after the roll. In addition, there is a new variable, roll.result, that lets you use the result of the roll in the tooltip parameter.