Contributed by mark karasek
[s: ] - only display to self
[g: ] - only display to GM
[w("name"): ] - only display to specific player
[st: ] - only display tooltip to self
[gt: ] - only display tooltip to GM

These can be combined, so, for example, the output of [s,g,st: ] would appear to yourself and the GM but not other players, and only you would have a tooltip for it.

There's also an additional change. If you receive a message that would be empty because, say, it consists of a bunch of rolls that aren't visible to you, the message is suppressed. This applies to [h: ] rolls too; sending "[h: d20]" would result in in no output instead of a blank message.