Contributed by cheesethulu: 1.  Corrects the bounding box for the line template.  It was being miscalculated when the line was drawn from right to left, which sometimes caused it to not be rendered in all of the chunks it belonged in.
2.  A line template that had its first and second vertices the same
would throw an NPE when getting its bounding box.  It should no longer
be possible to create line templates like that and ones in existing
campaigns will be deleted during load time optimization.  (fixes
problem mentioned in this thread:
http://forums.rptools.net/viewtopic.php?f=3&t=11183 )

3.  LineTemplateTool was calculating the quadrant of a line template
using a mix of a ZonePoint and a ScreenPoint, so when the origin of
the map wasn't at the top-left of the ZoneRenderer it would sometimes
get the direction wrong.

4.  Unfilled, erasing drawables no longer delete drawables contained
within their interior when load time optimization occurs.