From Craig (cwisniew@sf.net):
This patch allows for html in labels and prompts (the labels on left of the dialog).

If the prompt or the label in enclosed in <html></html> tag, then
it is rendered as html and links and tool tips work as the does clicking on links. Forms are not supported (nor do I see any reason to support them). The font for HTML rendered items is slightly different from normal labels so you will probably want to enclose even the non html prompts in <html></html> to get same font. I didn't want to always place it in a html control as they are quite a bit more memory intensive than a label.

Things that people will have to keep in mind while using this, the input dialog is modal, so it is not possible to pop up a frame/html dialog in front of it (they will always pop up behind). It is possible how ever to pop up another input dialog in front of it.